third_party/blink/web_tests/fast/dom/shadow/inner-scope-important-wins-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Outer scope !important rules wins with higher specificity (v0 bug compatibility).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(target).color is "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
